requirements.txt
astor==0.8.1
boutdata==0.2.1
boututils==0.2.1
cftime==1.6.4
clang-format==18.1.8
cupy-cuda12x==13.2.0
Cython==3.0.10
distlib==0.3.8
fastrlock==0.8.2
filelock==3.15.4
imageio==2.34.2
mpmath==1.3.0
natsort==8.4.0
netCDF4==1.7.1.post1
numpy==1.26.4
platformdirs==4.2.2
pyevtk==1.6.0
sympy==1.13.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@773bcc865af5d5a45b405c80faf6fcc3cc510d7d
virtualenv==20.26.3
zoidberg==0.1.1
